Tuple is now encapsulated better and added Identity matrix

+++ src/buildingBlocks/include/color.h
@@ -7,23 +7,23 @@ class Color
 {
 public:
    Color();
-   Color(float red, float green, float blue);
+   Color(double red, double green, double blue);
 
   bool operator==(const Color& rhs) const;
   Color operator+(const Color& rhs) const;
   Color operator-(const Color& rhs) const;
-  Color operator*(const float& rhs) const;
+  Color operator*(const double& rhs) const;
   Color operator*(const Color& rhs) const; // Hadamarad Product
 
   // Color setters and getters
-  const float& red() const;
-  void red(const float& val);
+  const double& red() const;
+  void red(const double& val);
 
-  const float& green() const;
-  void green(const float& val);
+  const double& green() const;
+  void green(const double& val);
 
-  const float& blue() const;
-  void blue(const float& val);
+  const double& blue() const;
+  void blue(const double& val);
 
   Tuple  _color;
 };
